+++ query_math.go
@@ -7,7 +7,7 @@ import (
 var (
 	// MinVal represents the smallest possible value RethinkDB can store
 	MinVal = constructRootTerm("MinVal", p.Term_MINVAL, []interface{}{}, map[string]interface{}{})
-	// MaxVal represents the smallest possible value RethinkDB can store
+	// MaxVal represents the largest possible value RethinkDB can store
 	MaxVal = constructRootTerm("MaxVal", p.Term_MAXVAL, []interface{}{}, map[string]interface{}{})
 )
 
@@ -183,4 +183,38 @@ func (t Term) Random(args ...interface{}) Term {
 	}
 
 	return constructMethodTerm(t, "Random", p.Term_RANDOM, args, opts)
+}
+
+// Round causes the input number to be rounded the given value to the nearest whole integer.
+func (t Term) Round(args ...interface{}) Term {
+	return constructMethodTerm(t, "Round", p.Term_ROUND, args, map[string]interface{}{})
+}
+
+// Round causes the input number to be rounded the given value to the nearest whole integer.
+func Round(args ...interface{}) Term {
+	return constructRootTerm("Round", p.Term_ROUND, args, map[string]interface{}{})
+}
+
+// Ceil rounds the given value up, returning the smallest integer value greater
+// than or equal to the given value (the value’s ceiling).
+func (t Term) Ceil(args ...interface{}) Term {
+	return constructMethodTerm(t, "Ceil", p.Term_CEIL, args, map[string]interface{}{})
+}
+
+// Ceil rounds the given value up, returning the smallest integer value greater
+// than or equal to the given value (the value’s ceiling).
+func Ceil(args ...interface{}) Term {
+	return constructRootTerm("Ceil", p.Term_CEIL, args, map[string]interface{}{})
+}
+
+// Floor rounds the given value down, returning the largest integer value less
+// than or equal to the given value (the value’s floor).
+func (t Term) Floor(args ...interface{}) Term {
+	return constructMethodTerm(t, "Floor", p.Term_FLOOR, args, map[string]interface{}{})
+}
+
+// Floor rounds the given value down, returning the largest integer value less
+// than or equal to the given value (the value’s floor).
+func Floor(args ...interface{}) Term {
+	return constructRootTerm("Floor", p.Term_FLOOR, args, map[string]interface{}{})
 }